🔧 (MigrationTools): centralize log path creation and improve logging
Refactor log path creation into a dedicated LogLocationService to centralize and simplify log path management. This change ensures consistent log path generation and usage across the application. Additionally, enhance logging by including log path information in log messages and enabling file size limits and rolling intervals for log files. This improves log management and helps in diagnosing issues by providing clear log locations.
